Refactor: Modularize FirstGroupPage
- FirstGroupPage Refactor:
  - Moved `FirstGroupPage.jsx` to a new folder structure: `src/pages/FirstGroupPage/`.
  - Split `FirstGroupPage.jsx` into smaller components:
    - `AddExpensesStep.jsx`, `AddGroupDetailsStep.jsx`, `AddParticipantsStep.jsx`, `LastStep.jsx`.
  - Created a `context` (`FirstGroupPageContext.jsx`) for managing stepper logic across all child components of `FirstGroupPage`.
  - Added a `useFirstGroupPageContext.js` hook for safer access to the context.
  - Organized reusable types in `types.js` and mock data in `mock-data.js`.

- New Reusable Components:
  - Added new reusable components in `src/components`:
    - `ExpensesList.jsx`, `GroupInfoWidget.jsx`, `SelectedParticipantsList.jsx`, and `Typography.jsx`.
    - The `Typography` component exports `Heading` and `BodyText` subcomponents.

- Other Improvements:
  - Fixed the z-index issue in `AppTopnav`.
  - Moved `formatCurrency` utility from `FirstGroupPage` to `src/lib/utils.js` for broader use.
