--- src/App.jsx
@@ -1,5 +1,0 @@
-function App() {
-  return <h1 className="text-blue-600 text-5xl font-bold">Hello world!</h1>
-}
-
-export default App
