+++ src/lib/constants.js
@@ -1,0 +1,27 @@
+export const CONTRIBUTION_WEIGHTS = [0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100]
+
+export const FIRST_GROUP_STEPS = [
+  {
+    id: "first",
+    title: "Add Group Details",
+    description: "Enter the basic information for your group, such as the group name and purpose.",
+  },
+  {
+    id: "second",
+    title: "Add Participants",
+    description:
+      "These participants will automatically be added to your friends list, making it easier to select them for future group expenses.",
+  },
+  {
+    id: "third",
+    title: "Add Expenses",
+    description:
+      "Start adding the expenses for the group. You can assign amounts and categorize the expenses for easy tracking.",
+  },
+  {
+    id: "last",
+    title: "You're Almost There! Finalize Your Group Setup",
+    description:
+      "Review the details and finalize the group setup. You'll be able to see insights like balances and statistics, and manage the expense group from the dashboard.",
+  },
+]
