+++ src/lib/utils.js
@@ -10,3 +10,12 @@ export function getInitials(name) {
   const initials = nameArray.map((word) => word.charAt(0).toUpperCase()).join("")
   return initials
 }
+
+export function formatCurrency(number) {
+  return new Intl.NumberFormat("en-US", {
+    style: "currency",
+    currency: "USD",
+    minimumFractionDigits: number % 1 === 0 ? 0 : 2,
+    maximumFractionDigits: 2,
+  }).format(number)
+}
