+++ src/pages/FirstGroupPage/components/types.js
@@ -1,0 +1,5 @@
+import PropTypes from "prop-types"
+
+export const StepsContentCommonTypes = {
+  includeActions: PropTypes.bool,
+}
